# openapi.yml
openapi: 3.0.3
info:
  title: English Learning App API
  description: API for managing users and audio files in the English learning application
  version: 1.2.1
  contact:
    name: English Learning App
    email: support@englishlearningapp.com
servers:
  - url: https://1bvgxdmq64.execute-api.us-east-1.amazonaws.com/dev
    description: Development server
security:
  - CognitoUserPool: []
paths:
...
  /audio/upload:
    post:
      operationId: generateUploadUrl
      summary: Generate presigned URL for audio upload
      tags: [Audio]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateUploadUrlRequest'
      responses:
        '200':
          description: Upload URL generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /audio/download:
    post:
      operationId: generateDownloadUrl
      summary: Generate presigned URL for audio download
      tags: [Audio]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDownloadUrlRequest'
      responses:
        '200':
          description: Download URL generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - file not accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /audio/metadata:
    post:
      operationId: getAudioMetadata
      summary: Get audio file metadata
      tags: [Audio]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetMetadataRequest'
      responses:
        '200':
          description: Metadata retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - file not accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /audio/files/{audioFileKey}:
    get:
      operationId: getAudioFileUrl
      summary: Get download URL for specific audio file
      tags: [Audio]
      parameters:
        - name: audioFileKey
          in: path
          required: true
          schema:
            type: string
          description: The audio file key (URL encoded)
      responses:
        '200':
          description: Download URL retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - file not accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteAudioFile
      summary: Delete audio file
      tags: [Audio]
      parameters:
        - name: audioFileKey
          in: path
          required: true
          schema:
            type: string
          description: The audio file key (URL encoded)
      responses:
        '200':
          description: Audio file deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - file not accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # User Endpoints
...
components:
  securitySchemes:
    CognitoUserPool:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito User Pool JWT token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
      required: [error]
...
    GenerateUploadUrlRequest:
      type: object
      properties:
        fileName:
          type: string
          description: Name of the audio file
          example: my-audio.mp3
        contentType:
          type: string
          description: MIME type of the audio file
          enum: [audio/mpeg, audio/wav, audio/m4a, audio/webm]
          example: audio/mpeg
        duration:
          type: number
          description: Duration in seconds (optional)
          minimum: 0
          example: 30
        fileSize:
          type: number
          description: File size in bytes (optional)
          minimum: 1
          maximum: 104857600
          example: 1048576
      required: [fileName, contentType]
    UploadUrlResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        audioFileKey:
          type: string
          description: Unique key for the uploaded file
          example: audio-files/user123/audio_20231201_143022.mp3
        uploadUrl:
          type: string
          format: uri
          description: Presigned URL for uploading the file
        expiresIn:
          type: number
          description: URL expiration time in seconds
          example: 3600
        maxFileSize:
          type: number
          description: Maximum allowed file size in bytes
          example: 104857600
        supportedFormats:
          type: array
          items:
            type: string
          description: List of supported audio formats
          example: [audio/mpeg, audio/wav, audio/m4a, audio/webm]
      required: [success, audioFileKey, uploadUrl, expiresIn, maxFileSize, supportedFormats]
    GenerateDownloadUrlRequest:
      type: object
      properties:
        audioFileKey:
          type: string
          description: The audio file key to download
          example: audio-files/user123/audio_20231201_143022.mp3
      required: [audioFileKey]
    DownloadUrlResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        downloadUrl:
          type: string
          format: uri
          description: Presigned URL for downloading the file
        expiresIn:
          type: number
          description: URL expiration time in seconds
          example: 3600
      required: [success, downloadUrl, expiresIn]
    GetMetadataRequest:
      type: object
      properties:
        audioFileKey:
          type: string
          description: The audio file key to get metadata for
          example: audio-files/user123/audio_20231201_143022.mp3
      required: [audioFileKey]
    MetadataResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        metadata:
          type: object
          properties:
            contentType:
              type: string
              example: audio/mpeg
            size:
              type: number
              example: 1048576
            lastModified:
              type: string
              format: date-time
            duration:
              type: number
              description: Duration in seconds
              example: 30
          required: [contentType, size, lastModified]
      required: [success, metadata]
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Audio file deleted successfully
      required: [success, message]

    # User Schemas
...